reformatted code with AStyle to match global indenting and formatting
- Fixed initializing joystick device under Linux
- Reorganized platform structure
- Fixed compiling on Linux
- Misc fixes
- Fixed sound system under Linux using ALSA

+++ src/botlib/be_ai_chat.c
@@ -128,7 +128,7 @@ typedef struct bot_synonym_s
 //list with synonyms
 typedef struct bot_synonymlist_s
 {
-    unsigned long int context;
+    unsigned int context;
     float totalweight;
     bot_synonym_t* firstsynonym;
     struct bot_synonymlist_s* next;
@@ -152,7 +152,7 @@ typedef struct bot_matchpiece_s
 //match template
 typedef struct bot_matchtemplate_s
 {
-    unsigned long int context;
+    unsigned int context;
     int type;
     int subtype;
     bot_matchpiece_t* first;
@@ -688,7 +688,7 @@ void BotDumpSynonymList( bot_synonymlist_t* synlist )
 bot_synonymlist_t* BotLoadSynonyms( char* filename )
 {
     int pass, size, contextlevel, numsynonyms;
-    unsigned long int context, contextstack[32];
+    unsigned int context, contextstack[32];
     char* ptr = NULL;
     source_t* source;
     token_t token;
@@ -868,7 +868,7 @@ bot_synonymlist_t* BotLoadSynonyms( char* filename )
 // Returns:					-
 // Changes Globals:		-
 //===========================================================================
-void BotReplaceSynonyms( char* string, unsigned long int context )
+void BotReplaceSynonyms( char* string, unsigned int context )
 {
     bot_synonymlist_t* syn;
     bot_synonym_t* synonym;
@@ -891,7 +891,7 @@ void BotReplaceSynonyms( char* string, unsigned long int context )
 // Returns:					-
 // Changes Globals:		-
 //===========================================================================
-void BotReplaceWeightedSynonyms( char* string, unsigned long int context )
+void BotReplaceWeightedSynonyms( char* string, unsigned int context )
 {
     bot_synonymlist_t* syn;
     bot_synonym_t* synonym, *replacement;
@@ -939,7 +939,7 @@ void BotReplaceWeightedSynonyms( char* string, unsigned long int context )
 // Returns:					-
 // Changes Globals:		-
 //===========================================================================
-void BotReplaceReplySynonyms( char* string, unsigned long int context )
+void BotReplaceReplySynonyms( char* string, unsigned int context )
 {
     char* str1, *str2, *replacement;
     bot_synonymlist_t* syn;
@@ -1651,7 +1651,7 @@ int StringsMatch( bot_matchpiece_t* pieces, bot_match_t* match )
 // Returns:					-
 // Changes Globals:		-
 //===========================================================================
-int BotFindMatch( char* str, bot_match_t* match, unsigned long int context )
+int BotFindMatch( char* str, bot_match_t* match, unsigned int context )
 {
     int i;
     bot_matchtemplate_t* ms;
@@ -2474,8 +2474,8 @@ int BotLoadChatFile( int chatstate, char* chatfile, char* chatname )
 // Returns:				-
 // Changes Globals:		-
 //===========================================================================
-int BotExpandChatMessage( char* outmessage, char* message, unsigned long mcontext,
-                          bot_matchvariable_t* variables, unsigned long vcontext, int reply )
+int BotExpandChatMessage( char* outmessage, char* message, unsigned mcontext,
+                          bot_matchvariable_t* variables, unsigned vcontext, int reply )
 {
     int num, len, i, expansion;
     char* outputbuf, *ptr, *msgptr;
@@ -2600,8 +2600,8 @@ int BotExpandChatMessage( char* outmessage, char* message, unsigned long mcontex
 // Returns:				-
 // Changes Globals:		-
 //===========================================================================
-void BotConstructChatMessage( bot_chatstate_t* chatstate, char* message, unsigned long mcontext,
-                              bot_matchvariable_t* variables, unsigned long vcontext, int reply )
+void BotConstructChatMessage( bot_chatstate_t* chatstate, char* message, unsigned mcontext,
+                              bot_matchvariable_t* variables, unsigned vcontext, int reply )
 {
     int i;
     char srcmessage[MAX_MESSAGE_SIZE];
