+++ src/cgame/cg_syscalls.c
@@ -42,13 +42,13 @@
 
 #include "cg_local.h"
 
-static int ( QDECL* syscall )( int arg, ... ) = ( int ( QDECL* )( int, ... ) ) - 1;
+static intptr_t ( QDECL* syscall )( intptr_t arg, ... ) = ( intptr_t ( QDECL* )( intptr_t, ... ) ) - 1;
 
 // TTimo: guarding
 #if defined( __MACOS__ )
 #pragma export on
 #endif
-void dllEntry( int ( QDECL*  syscallptr )( int arg, ... ) )
+void dllEntry( intptr_t ( QDECL*  syscallptr )( intptr_t arg, ... ) )
 {
 #if defined( __MACOS__ )
 #pragma export off
