+++ src/qcommon/vm_interpreted.c
@@ -345,7 +345,7 @@ locals from sp
 
 #define DEBUGSTR va( "%s%i", VM_Indent( vm ), opStack - stack )
 
-int VM_CallInterpreted( vm_t* vm, int* args )
+intptr_t VM_CallInterpreted( vm_t* vm, intptr_t* args )
 {
     int stack[MAX_STACK];
     int*     opStack;
@@ -524,7 +524,7 @@ nextInstruction2:
                 
                 src = ( int* )&image[ r0 & dataMask ];
                 dest = ( int* )&image[ r1 & dataMask ];
-                if( ( ( int )src | ( int )dest | count ) & 3 )
+                if( ( ( intptr_t )src | ( intptr_t )dest | count ) & 3 )
                 {
                     Com_Error( ERR_DROP, "OP_BLOCK_COPY not dword aligned" );
                 }
@@ -567,7 +567,7 @@ nextInstruction2:
                     *( int* )&image[ programStack + 4 ] = -1 - programCounter;
                     
 //VM_LogSyscalls( (int *)&image[ programStack + 4 ] );
-                    r = vm->systemCall( ( int* )&image[ programStack + 4 ] );
+                    r = vm->systemCall( ( intptr_t* )&image[ programStack + 4 ] );
                     
 #ifdef DEBUG_VM
                     // this is just our stack frame pointer, only needed
@@ -593,7 +593,7 @@ nextInstruction2:
                 }
                 goto nextInstruction;
                 
-                // push and pop are only needed for discarded or bad function return values
+            // push and pop are only needed for discarded or bad function return values
             case OP_PUSH:
                 opStack++;
                 goto nextInstruction;
@@ -650,12 +650,12 @@ nextInstruction2:
                 }
                 goto nextInstruction;
                 
-                /*
-                ===================================================================
-                BRANCHES
-                ===================================================================
-                */
-                
+            /*
+            ===================================================================
+            BRANCHES
+            ===================================================================
+            */
+            
             case OP_JUMP:
                 programCounter = r0;
                 programCounter = vm->instructionPointers[ programCounter ];
@@ -877,8 +877,8 @@ nextInstruction2:
                 }
                 
                 
-                //===================================================================
-                
+            //===================================================================
+            
             case OP_NEGI:
                 *opStack = -r0;
                 goto nextInstruction;
