+++ src/renderer/tr_backend.c
@@ -241,9 +241,9 @@ void GL_TexEnv( int env )
 ** This routine is responsible for setting the most commonly changed state
 ** in Q3.
 */
-void GL_State( unsigned long stateBits )
+void GL_State( unsigned int stateBits )
 {
-    unsigned long diff = stateBits ^ glState.glStateBits;
+    unsigned int diff = stateBits ^ glState.glStateBits;
     
     if( !diff )
     {
@@ -1827,7 +1827,7 @@ const void*  RB_SwapBuffers( const void* data )
     if( r_measureOverdraw->integer )
     {
         int i;
-        long sum = 0;
+        int sum = 0;
         unsigned char* stencilReadback;
         
         stencilReadback = ri.Hunk_AllocateTempMemory( glConfig.vidWidth * glConfig.vidHeight );
