+++ src/renderer/tr_image.c
@@ -156,7 +156,7 @@ return a hash value for the filename
 static long generateHashValue( const char* fname )
 {
     int i;
-    long hash;
+    int hash;
     char letter;
     
     hash = 0;
@@ -172,7 +172,7 @@ static long generateHashValue( const char* fname )
         {
             letter = '/';                   // damn path names
         }
-        hash += ( long )( letter ) * ( i + 119 );
+        hash += ( int )( letter ) * ( i + 119 );
         i++;
     }
     hash &= ( FILE_HASH_SIZE - 1 );
@@ -984,7 +984,7 @@ image_t* R_CreateImageExt( const char* name, const byte* pic, int width, int hei
 {
     image_t*     image;
     qboolean isLightmap = qfalse;
-    long hash;
+    int hash;
     qboolean noCompress = qfalse;
     
     if( strlen( name ) >= MAX_QPATH )
@@ -1108,20 +1108,20 @@ BMP LOADING
 typedef struct
 {
     char id[2];
-    unsigned long fileSize;
-    unsigned long reserved0;
-    unsigned long bitmapDataOffset;
-    unsigned long bitmapHeaderSize;
-    unsigned long width;
-    unsigned long height;
+    unsigned int fileSize;
+    unsigned int reserved0;
+    unsigned int bitmapDataOffset;
+    unsigned int bitmapHeaderSize;
+    unsigned int width;
+    unsigned int height;
     unsigned short planes;
     unsigned short bitsPerPixel;
-    unsigned long compression;
-    unsigned long bitmapDataSize;
-    unsigned long hRes;
-    unsigned long vRes;
-    unsigned long colors;
-    unsigned long importantColors;
+    unsigned int compression;
+    unsigned int bitmapDataSize;
+    unsigned int hRes;
+    unsigned int vRes;
+    unsigned int colors;
+    unsigned int importantColors;
     unsigned char palette[256][4];
 } BMPHeader_t;
 
@@ -1151,33 +1151,33 @@ static void LoadBMP( const char* name, byte** pic, int* width, int* height )
     
     bmpHeader.id[0] = *buf_p++;
     bmpHeader.id[1] = *buf_p++;
-    bmpHeader.fileSize = LittleLong( *( long* ) buf_p );
+    bmpHeader.fileSize = LittleLong( *( int* ) buf_p );
     buf_p += 4;
-    bmpHeader.reserved0 = LittleLong( *( long* ) buf_p );
+    bmpHeader.reserved0 = LittleLong( *( int* ) buf_p );
     buf_p += 4;
-    bmpHeader.bitmapDataOffset = LittleLong( *( long* ) buf_p );
+    bmpHeader.bitmapDataOffset = LittleLong( *( int* ) buf_p );
     buf_p += 4;
-    bmpHeader.bitmapHeaderSize = LittleLong( *( long* ) buf_p );
+    bmpHeader.bitmapHeaderSize = LittleLong( *( int* ) buf_p );
     buf_p += 4;
-    bmpHeader.width = LittleLong( *( long* ) buf_p );
+    bmpHeader.width = LittleLong( *( int* ) buf_p );
     buf_p += 4;
-    bmpHeader.height = LittleLong( *( long* ) buf_p );
+    bmpHeader.height = LittleLong( *( int* ) buf_p );
     buf_p += 4;
     bmpHeader.planes = LittleShort( *( short* ) buf_p );
     buf_p += 2;
     bmpHeader.bitsPerPixel = LittleShort( *( short* ) buf_p );
     buf_p += 2;
-    bmpHeader.compression = LittleLong( *( long* ) buf_p );
+    bmpHeader.compression = LittleLong( *( int* ) buf_p );
     buf_p += 4;
-    bmpHeader.bitmapDataSize = LittleLong( *( long* ) buf_p );
+    bmpHeader.bitmapDataSize = LittleLong( *( int* ) buf_p );
     buf_p += 4;
-    bmpHeader.hRes = LittleLong( *( long* ) buf_p );
+    bmpHeader.hRes = LittleLong( *( int* ) buf_p );
     buf_p += 4;
-    bmpHeader.vRes = LittleLong( *( long* ) buf_p );
+    bmpHeader.vRes = LittleLong( *( int* ) buf_p );
     buf_p += 4;
-    bmpHeader.colors = LittleLong( *( long* ) buf_p );
+    bmpHeader.colors = LittleLong( *( int* ) buf_p );
     buf_p += 4;
-    bmpHeader.importantColors = LittleLong( *( long* ) buf_p );
+    bmpHeader.importantColors = LittleLong( *( int* ) buf_p );
     buf_p += 4;
     
     memcpy( bmpHeader.palette, buf_p, sizeof( bmpHeader.palette ) );
@@ -1967,8 +1967,8 @@ jpeg_write_scanlines( j_compress_ptr cinfo, JSAMPARRAY scanlines,
     /* Call progress monitor hook if present */
     if( cinfo->progress != NULL )
     {
-        cinfo->progress->pass_counter = ( long ) cinfo->next_scanline;
-        cinfo->progress->pass_limit = ( long ) cinfo->image_height;
+        cinfo->progress->pass_counter = ( int ) cinfo->next_scanline;
+        cinfo->progress->pass_limit = ( int ) cinfo->image_height;
         ( *cinfo->progress->progress_monitor )( ( j_common_ptr ) cinfo );
     }
     
@@ -2221,7 +2221,7 @@ image_t* R_FindImageFileExt( const char* name, qboolean mipmap, qboolean allowPi
     image_t* image;
     int width, height;
     byte*    pic;
-    long hash;
+    int hash;
     
     if( !name )
     {
@@ -2905,7 +2905,7 @@ qhandle_t RE_GetShaderFromModel( qhandle_t modelid, int surfnum, int withlightma
             if( surf->shader->lightmapIndex > LIGHTMAP_NONE )
             {
                 image_t* image;
-                long hash;
+                int hash;
                 qboolean mip = qtrue;   // mip generation on by default
                 
                 // get mipmap info for original texture
